kernel/predicates: round every product at its source, and delete `rounded`
The static filter certifies a sign against Shewchuk's a-priori bound,
which is derived on the assumption that every multiply and add rounds
separately. A fused product does not merely make the estimate
platform-dependent — it invalidates the bound the certification rests on.
The package already knew that and carried a local `rounded` helper for
the products feeding the determinant, but its two `permanent` sums and
`dominantNormalAxis`'s cross product were left unrounded: 7 fused
multiply-add instructions in the arm64 build.

40 conversions now: 14 products the policy binds that carried none, and
the 26 `rounded(...)` calls respelled as the same explicit conversion so
the repo has ONE spelling of the rule (ADR-0064). The helper is deleted;
the package doc keeps its reasoning and names the guard that enforces it.

Measured: `make arm64-fma PKG=./kernel/predicates` 7 -> 0, and the
package's own tests pass under arm64 emulation.

Refs #3528

+++ kernel/predicates/queries.go
@@ -92,9 +92,9 @@ func SegmentPiercesTriangle(p, q, a, b, c [3]float64) bool {
 func dominantNormalAxis(a, b, c [3]float64) int {
 	ux, uy, uz := b[0]-a[0], b[1]-a[1], b[2]-a[2]
 	vx, vy, vz := c[0]-a[0], c[1]-a[1], c[2]-a[2]
-	nx := math.Abs(uy*vz - uz*vy)
-	ny := math.Abs(uz*vx - ux*vz)
-	nz := math.Abs(ux*vy - uy*vx)
+	nx := math.Abs(float64(uy*vz) - float64(uz*vy))
+	ny := math.Abs(float64(uz*vx) - float64(ux*vz))
+	nz := math.Abs(float64(ux*vy) - float64(uy*vx))
 	if nx >= ny && nx >= nz {
 		return 0
 	}
